+++ bricks/good_boilerplate/__brick__/{{project_name.snakeCase()}}/lib/main_dev.dart
@@ -10,12 +10,19 @@ void main() {
       baseApiUrl: 'https://dev-api.example.com',
       enableNetworkLogs: true,
       enableUpgradeCheck: true,
+      enableCodePush: false,
       gemini: GeminiConfig(
         apiKey: 'REPLACE_WITH_GEMINI_DEV_API_KEY',
         prompt: GeminiConfig.defaultPrompt,
       ),
       revenueCat: RevenueCatConfig(
         apiKey: 'REPLACE_WITH_REVENUECAT_DEV_API_KEY',
+        offeringIdentifier: 'default',
+        promoOfferingIdentifier: 'promo_offer',
+      ),
+      wiredash: WiredashConfig(
+        projectId: 'REPLACE_WITH_WIREDASH_DEV_PROJECT_ID',
+        secret: 'REPLACE_WITH_WIREDASH_DEV_SECRET',
       ),
       adMob: AdMobConfig(
         bannerAdUnitId: 'ca-app-pub-3940256099942544/6300978111',